ENH: Create converters.
I am not sure the best way to organize the converters. I want to satisfy these criteria

- each converter should be in its own file
- each filename should show the converter version
- (eventually) access converters dynamically through a dict, for a simple CLI interface

I haven't figured a good way to do this, while keeping the filenames as valid python module names. So I created custom import logic in converters.py which loads the converters, even though the file names are not valid module names. It's a gross violation of PEP8, but I think it improves readability here.